Fix check for existing flutter submodule
In my project I have a submodule named `flutter_custom_tabs`, which breaks the wrapper's check for the existing flutter submodule. This change [references](https://stackoverflow.com/a/10346824/3911006) the fact that the `.flutter` command in grep is using regex, meaning anyone else with submodules that match with the string `flutter` in it will fail silently here.