Update flutterw to git fetch before searching for tags
Found that this is required when the submodule has been fetched with a limited git depth,
as might be the case on CI servers.  Not doing this causes the tag search to fail.

Also updated any references to .flutter or the submodule name to use the variables
declared further up the script, for consistency.